docs/backups/duplicados_2_cleanup_20260411_140702/requirements (2).txt
streamlit
pandas
python-dotenv
sqlalchemy
pillow
pillow-heif
numpy
google-generativeai>=0.8.0
st-supabase-connection
supabase
bcrypt
streamlit-option-menu
extra-streamlit-components

